pyproject.toml
[project]
name = "bloggy"
version = "0.1.0"
description = "Ein einfacher Flask-Blog"
dependencies = [
    "flask",
    "markdown2",
    "python-frontmatter"
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-mock",
    "coverage"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
pythonpath = ["src"]